Added test framework.
