+++ DFC.App.JobProfileOverview.Tests.IntegrationTests.API/DFC.App.JobProfileOverview.Tests.IntegrationTests.API.csproj
@@ -23,6 +23,9 @@
     <None Update="appsettings.json">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <None Update="Resource\SOCCodeContentType.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
