--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Model/UpdateMoreInformationLink.cs
@@ -1,12 +1,0 @@
-﻿namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Model
-{
-    public class UpdateMoreInformationLink
-    {
-        public string Url { get; set; }
-        public string Title { get; set; }
-        public string Id { get; set; }
-        public string Text { get; set; }
-        public string JobProfileId { get; set; }
-        public string JobProfileTitle { get; set; }
-    }
-}
