--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Model/UpdateRegistration.cs
@@ -1,11 +1,0 @@
-﻿namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Model
-{
-    public class UpdateRegistration
-    {
-        public string Info { get; set; }
-        public string Id { get; set; }
-        public string Title { get; set; }
-        public string JobProfileId { get; set; }
-        public string JobProfileTitle { get; set; }
-    }
-}
