--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Model/UpdateRouteRequirement.cs
@@ -1,12 +1,0 @@
-﻿namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Model
-{
-    public class UpdateRouteRequirement
-    {
-        public string Url { get; set; }
-        public string Description { get; set; }
-        public string Id { get; set; }
-        public string Title { get; set; }
-        public string JobProfileId { get; set; }
-        public string JobProfileTitle { get; set; }
-    }
-}
