--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/HowToBecomeRouteEntry.json
@@ -1,8 +1,0 @@
-{
-  "RouteName": 0,
-  "EntryRequirements": [],
-  "MoreInformationLinks": [],
-  "RouteSubjects": "",
-  "FurtherRouteInformation": "",
-  "RouteRequirement": ""
-}
