--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/JobProfileCreateMessageBody.json
@@ -1,115 +1,0 @@
-{
-  "JobProfileId": "",
-  "Title": "Api Test Profile",
-  "DynamicTitlePrefix": "As Defined",
-  "AlternativeTitle": "Testing",
-  "Overview": "Test profile for API's",
-  "SocLevelTwo": "22",
-  "UrlName": "",
-  "DigitalSkillsLevel": "",
-  "Restrictions": [],
-  "OtherRequirements": "",
-  "CareerPathAndProgression": "",
-  "CourseKeywords": "'testing'",
-  "MinimumHours": 12.0,
-  "MaximumHours": 20.0,
-  "SalaryStarter": 12000.0,
-  "SalaryExperienced": 35000.0,
-  "WorkingPattern": [
-    {
-      "Id": "4784263f-add8-4e97-bfdc-de0ff5ad5b52",
-      "Title": "early mornings",
-      "Url": "early-mornings",
-      "Description": ""
-    }
-  ],
-  "WorkingPatternDetails": [],
-  "WorkingHoursDetails": [
-    {
-      "Id": "37d94efc-af05-4e54-8400-fe4b909dfb42",
-      "Title": "a week",
-      "Url": "a-week",
-      "Description": "Ismail  test"
-    }
-  ],
-  "HiddenAlternativeTitle": [],
-  "JobProfileSpecialism": [],
-  "IsImported": true,
-  "HowToBecomeData": {
-    "RouteEntries": [],
-    "FurtherInformation": {
-      "CareerTips": "",
-      "ProfessionalAndIndustryBodies": "",
-      "FurtherInformation": ""
-    },
-    "FurtherRoutes": {
-      "Work": "",
-      "Volunteering": "",
-      "DirectApplication": "",
-      "OtherRoutes": ""
-    },
-    "IntroText": "This is an automated introductory paragraph",
-    "Registrations": [
-      {
-        "Id": "",
-        "Title": "doesnt-matter",
-        "Info": "<p>Initial text</p>"
-      }
-    ]
-  },
-  "WhatYouWillDoData": {
-    "DailyTasks": "",
-    "Locations": [],
-    "Uniforms": [],
-    "Environments": [],
-    "Introduction": ""
-  },
-  "SocCodeData": {
-    "Id": "0a26f2dd-22fd-468d-b64c-9413bd3674fd",
-    "SOCCode": "2211A",
-    "Description": "Description of SOC Code for GP - Karl",
-    "ONetOccupationalCode": "29-1063.00",
-    "UrlName": "2211a",
-    "ApprenticeshipFramework": [
-      {
-        "Id": "a6718d26-9db8-4894-8394-b0facd72422d",
-        "Title": "Activity Leadership",
-        "Url": "460",
-        "Description": ""
-      },
-      {
-        "Id": "68de9a73-8a0e-41d2-98d1-d1a5bc0f38b9",
-        "Title": "Accounting ",
-        "Url": "454",
-        "Description": ""
-      }
-    ],
-    "ApprenticeshipStandards": [
-      {
-        "Id": "fe2a4688-b599-436b-b968-f5f3b2305a94",
-        "Title": "Able Seafarer (Deck) (Level 2)",
-        "Url": "34",
-        "Description": ""
-      },
-      {
-        "Id": "8736f15d-8dfd-483c-bfa2-3eab0e2cf318",
-        "Title": "Abattoir Worker (Level 2)",
-        "Url": "274",
-        "Description": ""
-      }
-    ]
-  },
-  "RelatedCareersData": [],
-  "SocSkillsMatrixData": [],
-  "JobProfileCategories": [
-    {
-      "Id": "fa4eea9e-b03c-414f-98bb-43b64edf46eb",
-      "Name": "emergency-and-uniform-services",
-      "Title": "Emergency and uniform services"
-    }
-  ],
-  "LastModified": "2020-01-10T08:48:43.37Z",
-  "CanonicalName": "",
-  "WidgetContentTitle": "",
-  "IncludeInSitemap": true
-}
