--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/JobProfileDeleteMessageBody.json
@@ -1,34 +1,0 @@
-{
-  "JobProfileId": "",
-  "Title": "Api Test Profile",
-  "DynamicTitlePrefix": null,
-  "AlternativeTitle": null,
-  "Overview": null,
-  "SocLevelTwo": null,
-  "UrlName": null,
-  "DigitalSkillsLevel": null,
-  "Restrictions": null,
-  "OtherRequirements": null,
-  "CareerPathAndProgression": null,
-  "CourseKeywords": null,
-  "MinimumHours": null,
-  "MaximumHours": null,
-  "SalaryStarter": null,
-  "SalaryExperienced": null,
-  "WorkingPattern": null,
-  "WorkingPatternDetails": null,
-  "WorkingHoursDetails": null,
-  "HiddenAlternativeTitle": null,
-  "JobProfileSpecialism": null,
-  "IsImported": null,
-  "HowToBecomeData": null,
-  "WhatYouWillDoData": null,
-  "SocCodeData": null,
-  "RelatedCareersData": null,
-  "SocSkillsMatrixData": null,
-  "JobProfileCategories": null,
-  "LastModified": "0001-01-01T00:00:00",
-  "CanonicalName": null,
-  "WidgetContentTitle": null,
-  "IncludeInSitemap": false
-}
