--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/UpdateEntryRequirement.json
@@ -1,7 +1,0 @@
-{
-  "Info": "",
-  "Id": "",
-  "Title": "doesnt-matter",
-  "JobProfileId": "",
-  "JobProfileTitle": "Api Test Profile"
-}
