--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/UpdateMoreInformationLinks.json
@@ -1,8 +1,0 @@
-{
-  "Text": "",
-  "Url": "https://does-not-matter.co.uk/",
-  "Id": "",
-  "Title": "doesnt-matter",
-  "JobProfileId": "",
-  "JobProfileTitle": "Api Test Profile"
-}
