--- DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Resource/UpdateRouteRequirement.json
@@ -1,8 +1,0 @@
-{
-  "Url": "does-not-matter",
-  "Description": "does-not-matter",
-  "Id": "",
-  "Title": "",
-  "JobProfileId": "",
-  "JobProfileTitle": "Api Test Profile"
-}
