+++ DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Support/Interface/IJobProfileOverviewSupport.cs
@@ -1,0 +1,9 @@
+﻿using DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Model;
+
+namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Support.Interface
+{
+    interface IJobProfileOverviewSupport
+    {
+        SOCCodeContentType GenerateSOCCodeContentType();
+    }
+}
