+++ DFC.App.JobProfileOverview.Tests.IntegrationTests.API/Test/Test.cs
@@ -1,6 +1,8 @@
+using DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Model;
 using DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Support;
 using NUnit.Framework;
 using System.Threading.Tasks;
+using static DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Support.EnumLibrary;
 
 namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Test
 {
@@ -9,6 +11,9 @@ namespace DFC.App.JobProfileOverview.Tests.IntegrationTests.API.Test
         [Test]
         public async Task ProofOfConcept()
         {
+            SOCCodeContentType socCodeContentType = CommonAction.GenerateSOCCodeContentType();
+            byte[] messageBody = CommonAction.ConvertObjectToByteArray(socCodeContentType);
+            CommonAction.CreateServiceBusMessage(socCodeContentType.Id, messageBody, ContentType.HTML, ActionType.Published, CType.JobProfileSoc);
 
         }
     }
